basic email locally passes
